fix LFO rates slower than 1/4 note, slew aftertouch

+++ Instrument/Shared/LFOKernel.hpp
@@ -140,7 +140,9 @@ public:
         transportState = state;
         
         if (sync) {
-            int startOfCycle = (transportState->currentBeatPosition / lfoRates[syncRateIndex].beatsPerCycle) * lfoRates[syncRateIndex].beatsPerCycle;
+            
+            int startOfCycle = ((int) (transportState->currentBeatPosition / lfoRates[syncRateIndex].beatsPerCycle)) * lfoRates[syncRateIndex].beatsPerCycle;
+            
             double position = transportState->currentBeatPosition - (double) startOfCycle;
             
             double lfoPhase = position / lfoRates[syncRateIndex].beatFrequency;
